+++ csharp.xUnit/GildedRoseTests/GildedRoseTest.cs
@@ -6,59 +6,6 @@ namespace GildedRoseTests;
 public class GildedRoseTest
 {
     [Fact]
-    public void NormalItem_DayPassed_ShouldDecreaseSellIn()
-    {
-        // Given
-        var sellIn = 0;
-        var item = Any.NormalItem(sellIn:sellIn);
-
-        // When
-        GildedRose.HandleNormalItem(item);
-
-        // Then
-        Assert.Equal(sellIn - 1, item.SellIn);
-    }
-
-    [Fact]
-    public void NormalItem_DayPassed_QualityDecreased()
-    {
-        // Given
-        var sellIn = 1;
-        var quality = 10;
-        var item = Any.NormalItem(sellIn:sellIn, quality:quality);
-
-        // When
-        GildedRose.HandleNormalItem(item);
-
-        // Then
-        Assert.Equal(quality - 1, item.Quality);
-    }
-
-    [Fact]
-    public void NormalItem_DayPassed_QualityNotBelowZero()
-    {
-        // Given
-        var item = Any.NormalItem();
-
-        // When
-        new GildedRose([item]).UpdateQuality();
-
-        // Then
-        Assert.Equal(0, item.Quality);
-    }
-
-    [Fact]
-    public void NormalItem_SellInPassed_QualityTwiceAsFast()
-    {
-        var quality = 10;
-        var item = Any.NormalItem(quality: quality);
-
-        new GildedRose([item]).UpdateQuality();
-
-        Assert.Equal(quality - 2, item.Quality);
-    }
-
-    [Fact]
     public void AgedBrie_Fresh_ShouldIncreaseInQuality()
     {
         var quality = 0;
